# Warning: the following style issues were flagged by pep8: #	src/parser/matlab/scope.py:20:1: E302 expected 2 blank lines, found 1 #	src/parser/matlab/scope.py:52:72: E203 whitespace before ':' #	src/parser/matlab/mat_parse.py:4:9: W291 trailing whitespace #	src/parser/matlab/mat_parse.py:27:1: E901 TokenError: EOF in multi-line statement # Please consider fixing the issues listed above before committing. # If you run into a situation where pep8 flags a line but it really # should not, or you prefer to leave that line as-is, add the # comment '# noqa' at the end of that line to make pep8 ignore it. New class for tracking scopes.
Rather than attaching a bunch of attributes directly to ParseResults
structures, we can gather them into a class object.  This file
implements that class, named "Scope" for a lack of better names.  It
provides some basic methods to manipulate the scope.

The intention is to attach an instance of a Scope object to a
ParseResults structure under an attribute named "scope", so that the
stuctures can be accessed as, e.g., "pr.scope".