Massive overhaul. This is vastly different from the previous iteration, with so many changes it is impossible to enumerate them in detail.  Here are the highlights:
1. Many parsing bugs fixed.

2. The new approach no longer uses the "tags" from the previous one,
and instead uses dictionaries to associate parse results with
interpretations.  A long comment near the beginning of the file
explains the principles.

3. There is a new function to print an annotated interpretation of the
parse results.  The output is a bit grungy but does serve the purpose
of letting humans check whether it parsed an input correctly.

4. New function make_formula() can take an expression (a subset of the
parse results that MatlabGrammar can produce) and return a string
representation.

5. Many more comments inside the file try to explain what's going on.