Vastly rewritten and improved. Highlight of the changes:
1. Fixes many bugs reported by Harold Gomez.

2. Works with the new version of MatlabGrammar, with its
dictionary-based approach to annotating parse results.

3. Now offers a flag to indicate whether variables should be made as
SBML species or SBML parameters.

4. Changes the base name of the variable used for the
species/parameters; now it uses the name of the variable inside the
matrix to which the results of the ode* calls are assigned.  In other
words, if the call is [t x] = ode45(@f, tspan, xinit), the variables
will have a root name of "x_".  (It will not be the name of the
variable inside the function, which is what the previous version did.)
The new behavior seems more in line with what people might expect.

5. Many more comments inside the file try to explain what's going on.