Implemented simple type inference for variables.
It will now take note of assignments and variable definitions, and
remember the names of those that are likely to be variables rather
than functions.  Then, upon encountering 'matrix or function', if
it's seen the name before and the name has been categorized as a
variable, it will change the object to 'matrix'.
  This is used to relabel matrices when possible.


Former-commit-id: 88fcfeaacacd09a96add462511ef11c27bb33ef5
Former-commit-id: a664ff9aa7f672e66b3c9f9b06f5d9eee7661668