Switched to using the word "index" intead of "subscript".
Reading the MATLAB documentation, it seems that they prefer to use the
term "subscript" when referring to the positional index terms of
matrices and arrays.  Probably a trivial issue, but avoiding the word
"index" does have the additional advantage that it's less likely to
cause confusion with respect to using linear indexing with arrays, so
I went ahead and changed it.


Former-commit-id: 4454847d944aa26b109b8bf6c69e012c8c78f8c9
Former-commit-id: a66f488781849e76e97ee52cace4b9cc96d4d49f