Switched to using the word "index" intead of "subscript".
Reading the MATLAB documentation, it seems that they prefer to use the
term "subscript" when referring to the positional index terms of
matrices and arrays.  Probably a trivial issue, but avoiding the word
"index" does have the additional advantage that it's less likely to
cause confusion with respect to using linear indexing with arrays, so
I went ahead and changed it.


Former-commit-id: 4183bcff0762d17e4a781e26b073a44a3c5c37a8
Former-commit-id: 640fd5cc6f8ca6aa881672ed69103c31141522ef