Renamed 'matrix' to 'array'.
I tended to think of 'matrix' as a generic term, potentially referring
to something with more than 2 dimensions.  In the parser, I also used
that term instead of the term 'array' because there's another Matlab
data type, the cell array, and wanted to make sure they weren't
confused.  However, it turns out this is not conventional, and in the
Matlab docs, they do tend to use the word 'matrix' to refer to 2-D
arrays only.  More than 2-D, and they call it an 'array'.  This
matters little to us, but might help future users/developers, so after
some discussion with the rest of the team, I'm going ahead and
renaming things to use the word 'array'.

Note that there is no easy way right now to determine the number of
dimensions of an object, so we can't distinguish vectors from (true)
matrices from arrays -- the parser tags anything that's not scalar as
'array', whether it has 1, 2 or more dimensions.


Former-commit-id: 09fe85dc27aab45570783b557dd98fb2d68f2a5f
Former-commit-id: a2f03cde9da978ddc613d3f636b2c263493c26ef