Added support for indirect ODE functions.
This now supports the case where the call to the matlab ode* function
uses a variable as the first argument, instead of a function handle.
It will look up the value of the variable to try to see if that's the
function handle.  If not, it gives up.

This supports a use case found in one of Matthias Koenig's models.


Former-commit-id: b1b2679a0caaa75dd4ded9d0514cdef676c202cc
Former-commit-id: 7b29b8a340f4353a3f8c831badb6ec3d1eb136e7