Bug fixes + better support for anonymous functions.
1. Fixed bug in _remember_type() caused by an incorrectly written
statement.

2. Moved function terminal() out of make_key() and made it a static
function on its own, so that it could be called from other places.

3. It now tries to disambiguate 'array or function' types inside
anonymous functions.  It uses the principle that the parameter names
in the anonymous function definition are variables, and thus if they
are referenced inside the function's body and the reference uses
'array or function' syntax, they are probably 'array' references and
not function calls.


Former-commit-id: 8a7b0e88ca9481c8275aa9483612d32c16e854bf
Former-commit-id: d80741b21de5f03d375124dc6f586a51acff1009