Made the tests work from any directory.
This required a change not only to the way directories are handled,
but also to how it creates the parser object.  The MatlabParser object
has to be recreated every time a new case is run, which is not what it
was doing before.  If it is not recreated every time a new case is
run, then it will retain memory of entities from one run to the next,
and this means that identifiers can be incorrectly tagged because
they're remembered from a previous parsing run.


Former-commit-id: 501f76fe32b92d296fa57842d42e5899ac780cbf
Former-commit-id: 3fd067c12fa0ea6d775e061599506551563ad73f