Made the tests work from any directory.
This required a change not only to the way directories are handled,
but also to how it creates the parser object.  The MatlabParser object
has to be recreated every time a new case is run, which is not what it
was doing before.  If it is not recreated every time a new case is
run, then it will retain memory of entities from one run to the next,
and this means that identifiers can be incorrectly tagged because
they're remembered from a previous parsing run.


Former-commit-id: a08af7c0c2b0cb5ab5ab4b89ea05021682f3b038
Former-commit-id: 4e5658210daf118b96dce751f8697320091dce8b