Rewrote to provide a better output representation.
Instead of returning an annotated, nested ParseResults, the new output
consists of lists of new object classes derived from MatlabNode.  In
addition, the value returned by MatlabGrammar.parse_string(...) is now
a MatlabContext, and it contains a field ("nodes") containing the
MatlabNode-based representation of the input.  A summary of the
new scheme is available in the comments in the file.  The object
classes are defined in the file matlab.py.

Also fixed a parsing grammar bug or two while I was at it.


Former-commit-id: 95affc96beee8b13a9ea8838ad0fd9be4d6d6e70
Former-commit-id: a234f1c1b457a2ef02a28d1a4346d810008da3d1