Several bug fixes and a minor improvement:
1) The code in make_formula did not apply the transformation function
to the arguments of function calls, which led to the problem that
expressions such as "ceil(k * x(1))" could not be rewritten to
"ceil(k * x_1)" by the converter.  Fixed.

2) Assignments were being stored before nodes were being processed in
_generate_nodes_and_contexts(), which caused assignment values
retrieved in the converter to be un-processed assignment values.  So,
simplifications applied to expressions were not stored

3) Function calls were being stored before nodes were processed, which
meant that changes to the arguments were not being stored in the
context.assignments dictionary.  The result was a divergent
representation between the node structure and the stored values in the
context object.  Fixed.

The improvement is that unary minus numbers are simplified to a Number
object with a value of a negative number, instead of being left as a
list consisting of a UnaryOp and a Number.  This makes it slightly
easier for the converter to generate parameters with simple numbers
rather than initial assignments, although in truth I now don't know if
that was worth it because Sarah's work on the number evaluator may
make this part moot.  Nevertheless, I left it in partly as a reminder
that other such simplifications may be possible.

(cherry picked from commit 38af38496ef6e83b29dc3317c6f8c513deb4b5c5 [formerly b0224f104d31fde22aeecc0958cd928ec95ba816] [formerly f4dc3a8d9cc32bee22687e919837645a03137cf4])


Former-commit-id: d1d2be0319d41b96ee2b7bd9480fd28cc0a51e25
Former-commit-id: d617762896ca6eb63c88c3044e4adfa6adf89e6a