More bugs fixed.
1) create_raterule_model() ignored the creation of parameters that
merely referred to other parameters. Fixed.

2) create_raterule_model() ignored the creation of parameters whose
values were bare function calls (i.e., not inside expressions).  Fixed.

3) Some models refer to time, sometimes with the variable name "t".
Added code to detect the use of an identifier "t" and create an
assignment rule for it that refers to time.  Not sure if that's the best
approach but it was a quick way to deal with it.


Former-commit-id: 6ad50e7072b54e9158f90cf3c45fbed89ae7503b
Former-commit-id: efeb347ef46235d98d6457597011a7a8a917ab62