Fixed brokenness in the matlab rewriting code.
The matlab rewriting code was not traversing everything.  The
assignments and calls dictionaries of the context object were getting
updated, but the nodes for the full tree were not, which meant there
were inconsistencies in the representations used in different parts of
the code.

To fix this, I finally implemented proper visitor pattern support in
matlab.py and used the visitor pattern here. (This turned out to be
much easier than I feared, and now I wonder why I didn't do that long
ago.)  Anyway, the result works better, traverses everythihng, and is
easier to understand.

To make it work, I had to change how globals were handled once more
time because the previous implementation somehow did not work properly
inside subclasses.


Former-commit-id: 6f1f67a939d7b12740a8ac0275a0bb5469338ac8
Former-commit-id: 378baeee0f027760135660ae36b53906826b3586