Multiple changes:
1) Make the default species name be after the parameter handed to the
ode function, rather than the output variable assigned in the
function.  This seems to be the more common assumption among users.  A
new command-line flag (-l) allows users to get the old behavior.

2) Look through comments in the file to try to infer the names of the
species variables.  This is simple and not at all general right now, but
handles the way that the Mount Sinai group has been commenting their
MATLAB code.

3) Fixed some very minor issues (probably merge problems) from Sarah's
last edits to the 'new-convert-to-ode' branch.

4) Changed some variables in XPP-oriented functions that referred to
"model" to use the variable name "xpp_variables" instead, because that's
what the calling code was using and I was getting confused about what
was going on when the parameter was named "model".


Former-commit-id: 81978d831977059247513fc39a335e59977d0231
Former-commit-id: 4c6c6a0d7af47fc03a484e2a52d6670c696e0954