Layout changes + fix for unicode issue
1) Tweaked the layout to look better on OSX.  Did not test on other
platforms yet.

2) Changed dialogs to use standard dialogs instead of the "generic"
wxPython dialog library, so that the dialogs look more native on OSX.

3) Fixed a problem involving the matlab input text getting changed to
unicode because the text was read into a text widget and then read out
again.  The fix involves storing the text read from the input file into
a variable and then working off that instead of reading it again from
the text widget.

4) Fixed conversion to work with new post-processing step.


Former-commit-id: 4b3c0ebd84b636fd46d5f30cd0e0967282d9d9b9
Former-commit-id: a7c5bc57477fe8c08acd7cbea0be55de9fb8c3d6