Added a flag to indicate the topmost context.
Being able to detect the top-level context turns out to be necessary
in a couple of places in grammar.py.  Previously I was doing it by
testing for the parent of the parent being null, but that's kind of
stupid.  It's better to have a flag that indicates when a context is
the top one.