Changed behavior of rewrite_recognized_matlab().
A wrinkle in how visitors like MatlabRewriter work is that they don't
have any context when they encounter a node, so it's difficult to
exclude certain cases.  Here we have a case where we want to detect
the presence of an identifier named 't' in the input, but we don't
want to do it if 't' is a parameter in a function definition.
However, there's no easy way to do that with a visitor approach.

To partly work around it, I changed rewrite_recognized_matlab() so
that it doesn't descend across all nodes in a file, but instead, just
the ones in assignments and function calls.  This actually is okay;
for our purposes, we *probably* don't care about the use of 't' in
situations other than assignments and function calls, although maybe
we simply haven't encountered the right examples.

With this change, the converter will not count 't' if it appears in a
function definition.  It will still count 't' if it shows up in an
anonymous function definition, though.  The consequence of this is
that the converter will occasionally introduce 't' (and the SBML
assignment rule for it) for models where 't' is not actualy needed.
For our current test cases, this happens in just one test case.