Enhanced the MatlabNodeVisitor.
The enhanced version features two main improvements.  First, it
automatically descents in the substructures of those classes that have
structure, such as the rows of arrays and the arguments to function
calls, so that subclasses do not have to implement this explicitly.
This removes a potential source of errors compared to the previous
version of MatlabNodeVisitor: now anyone who uses it to implement
visitor classes will not need to worry about putting in the right code
to make that happen (they did before).  Second, MatlabNodeVisitor now
has an explicit method for handling lists of nodes, so that subclasses
can implement special-casse handling of lists if they wish.

Both of these features are used in grammar.py now.