Rewrote to use the Visitor Pattern more thoroughly.
This now abstracts away the traversal of trees of ParseResults and
MatlabNode using new or improved visitor classes defined in matlab.py
and grammar_utils.py.  This hopefully simplifies the code and makes it
more comprehendable to anyone trying to figure out how it works; it
also hopefully lays the groundwork for more easily adding kind
analysis of MATLAB identifiers.