Fixed bug involving import statements.
This fixes the issue reported by Moretti Sebastien on August 1, where
using an absolute path to the scripts in moccasin/interfaces would
fail with a Python import error.  I changed the scripts to get the
current path to the .py file being executed, then add to sys.path a
path relative to that.