Improvement to recognizing functions.
This change allows the parser to recognize a function as a function
even when it is defined later in a file, so that it can handle calls
to functions from within other functions in the same file.  It is not
strictly valid for Matlab, but for now, it is a simple way to handle a
case that we didn't handle before.