Changed the option letters and added a new option.
The new option (using -d, which was previously used for something
else) makes the parser print all the grammar objects as its parsing
the input.  This is useful for debugging the grammar, although the
output is huge so you never want to use it for anything except a few
lines of input put in a simple test file.