Got rid of ParseResultsVisitor, added setVar(), changed nonempty_dict().
ParseResultsVisitor needed to be too specialized to NodeTransformer in
grammar.py, and since it was only used in that one class, it seemed
better to simply move the thing over there and get rid of this class.

Renamed nonempty_dict() to empty_dict() and changed the places it was
called to use "not empty_dict()".