Adjusted command-style function recognition.
Turns out colon needs to be added or some of the forms in test case
valid_201.m fail to parse correctly.  However, it's still wrong,
because it will parse "a - b" incorrectly when it appears as a
standalone statement.  It's going to be impossible to fix via
regexp-style recognition; it will need special processing to fix the
rest of the way.