Massive changes to fix multiple intertwined issues.
Parsing was still wrong for many situations.  After more hours than I
can count, I think I finally have it to where nearly everything is
correctly tokenized.  It now handles things it never did before, such
as the difference between nested functions and local functions, and
many more crazy MATLAB syntax beahviors.  My favorite example of the
latter is the following, if "a" is an array of one element:

  a(1)      => one item, the value inside the array "a"
  a (1)     => one item, the value inside the array "a"
  [a (1)]   => an array of TWO items, the value of a and 1
  [(a (1))] => an array of ONE item, a(1)

I thought I had it right before, but it turns out that it was still
incorrectly handled.

In addition, this now does more powerful kind analysis to try to
resolve cases of ambiguous references.  Among the things it does is
look around at how functions are called elsewhere in the file to
figure out if the parameters appear to be intended to be variables or
functions.  This helps handle cases such as

  function y = foo(x)
     y = x(1)   % Is this a function call, or an array reference?
  end

The level of intelligence in the inference system is still limited,
but it is better than it was before and really helps to disambiguate
our most common use cases.

For cases where something couldn't be inferred, it returns an object
of the new class "Ambiguous".  The previous ArrayOrFunCall object
class is gone.  The new Ambiguous object will have one of 3 possible
values for its 'arg' attribute:

1) the value None => the expression had no parentheses
2) the value []   => the expression had "()" for the arguments/subscripts
3) a list         => the expression had nonempty argument/subcripts

Case #1 is something like "foo" where we don't know whether "foo" is a
function being called without arguments or a variable being
referenced.  Case #2 is something like "foo()" where we don't know
whether "foo" is a function or an array variable.  Case #3 is similar
to #2, but with arguments or subscripts.

Finally, a lot of test cases were actually incorrectly parsed
previously, and other test cases had to change to account for the
demise of ArrayOrFunCall and the introduction of Ambiguous.  So, along
with this check-in, a lot of test case changes are checked in.