Changed handling of name of topmost context.
The name of the topmost context is now be the name of the first
function defined in the file, instead of the previous useless string
(which was "(topmost context)").  In combination with the Boolean
attribute 'topmost', this makes it possible to determined which one of
the function definitions in the 'functions' dictionary is the first
function defined in the file.  (Python dictionaries are unordered,
hence the need for something explicit.)