Recursive visitor algorithm had a subtle but critical flaw.
The approach of doing depth-first visits was almost correct, but
screwed up in the case of function definitions.  There, we need to
invoke the user's visit_FunDef() function *first* because it may need
to do things like change contexts before the body of the function
definition is processed.  What the visit() method on MatlabNodeVisitor
did instead, though, is visit the attributes including the body before
invoking the user's visit_FunDef() method.  This made correct context
switching impossible.

The new version always invokes the user's visit_foo() method if one
exists, and does a default walk of the attributes if it doesn't.
Also, if a visit_foo() method exists, then MatlabNodeVisitor.visit()
no longer visits the child attributes of the object like it did
before: the user's visit_foo() method must take care of doing that as
well.  This requires more work in the calling code, and is more risky
because people might forget to implement that, but it's the only way
to give the caller the power to do things in the order it needs them
to be done.  Otherwise, if visit() visits child attributes even if a
visit_foo() exists, then the user's visit_foo() method will never be
able to override the behavior if needs to.