Several intertwined changes.
1) MatlabNode objects can now be used as dict keys directly, instead
of having to use make_key() to create strings that are then used as
keys.  I discovered that in Python, you can use arbitrary objects as
dictionary keys if you define a __hash__() method on the object class.
I moved make_key() to matlab.py and used it as the basis of
MatlabNode's __hash__() implementation, thus making it possible to use
MatlabNode objects directly as dict keys.

2) With this change, a variety of things in the code here had to
change, especially with respect to methods like _save_type() and
_get_type().

3) This made me discover a bug in the way that contexts were being
tracked.  The bug was due to MatlabNode.visit()'s approach of doing
depth-first traversal of child attributes even if a derived class such
as NodeTransformer defined a visit_foo() method.  This was a problem
for function definitions: the body of the function definitions were
processed first before the visit_FunCall() method itself was called,
whereas visit_FunCall() really needs to be called first because it
needs to set up a new context for the body.  This required changing
the algorithm for MatlabNode.visit(), which had a knock-on effect on
the implementation of derived classes.  Now, all the derived classes
also have to be the ones responsible for visiting child attributes of
a class Foo if there exists a visit_Foo() method for Foo.