Several corrections
Getting converter.py running again has led to the discovery of a few
errors, particlarly with respect to context handling as well as
make_formula().  Also, in order to implement a more thorough visitor
algorithm in MatlabNodeVisitor, added some initialization code which
in turn requires subclasses of MatlabNodeVisitor to invoke the super()
method.