Implemented automatic superclass visiting in MatlabNodeVisitor.
Implemented support for writing visitor functions that work on
abstract superclasses of MatlabNode classes.  Now, subclasses of
MatlabNodeVisitor can implement functions like visit_Expression() and
they will be invoked on subclasses of Expression when an explicit
class visitor is not provided.

Also added a comment illustrating the class hierarchy at the top of
the file, because I got tired of looking in grammar.py for it.