Fixed arguments to create_raterule_model().
At some point I must have added or changed the arguments to
create_raterule_model() in converter.py, and this caused the behavior
to be different.  Now the calls use keyword arguments to prevent
ambiguity and accidentally using the wrong arg values.

Also increased the recursion depth limit.  Not sure that was needed,
but I was having trouble debugging something and suspected it might be
related.  Not convinced in the end, but leaving the setting anyway to
be safe.