Finally squashed pernicious, randomly-occurring parsing bug.
The problem was that one of the steps retrieved a value from a
dictionary, but it assumed there was only one value.  Turns out the
grammar definition for "ambiguous id" resulted in a ParseResults
construct getting two tags: "ambiguous id" and "identifier".
Consequently, sometimes the code would retrieve the value "identifier"
instead of "ambiguous id", and get a different result.

The correct fix was to change the definition of the grammar to avoid
getting "identifier" in addition to "ambiguous id".  I also added a
check to catch possible future occurrences of having two tags for a
given construct, so that it never happens again.  And just for good
measure, also added a sort step in first_key() so that it would
produce consistent results from run-to-run.