Egregious hack to solve python 2.7 py.test character encoding issue.
Py.test in python 2.7 seems to encode or decode characters somehow
differently from python 3.4, and the result is that non-ascii
characters get quoted when they are printed to stdout by the
underlying program being tested.  I tried to find the right solution
but couldn't, and gave up, and used a hack to work around the specific
problem so that at least I can get Travis tests to pass for now.  The
problem does not manifest when running outside of py.test, so it's not
a truly hideous hack, but it's not right either.