Speed up parsing by relaxing one of the grammar rules
The fact that a MATLAB file can contain either function definitions
with ends or without ends, but not mixed, led to a grammar that had an
OR condition, and this turned out to drop parsing speed by half.
Removing the disjunction speeds things up by almost 2x, at the cost of
not detecting certain invalid MATLAB files.