Fix for issue #52
The code that tries to adjust numbers for Biocham and XPP seems to
cause havoc when a number is very large.  I put in a new condition
that stops the procedure when a number is large.  The value chosen is
arbitrary (1e20) because I don't know what the real threshould should
be, but 1e20 is such a large number in biological contexts that
anything larger is probably a fake number anyway.