Fix for issue #53
Biocham currently (as of today, 2018-05-11) seems to mistranslate
unary negation, at least in some situations.  I added a hack that
replaces expressions of the form "-x" with "-1 * x", which works
better when passed to Biocham.  This is hopefully temporary, until
Biocham has a chance to address whatever the underlying issue is.