Stop error if you type into the file path box
Right now it simply ignores anything typed into the field. I can't
figure out how to let the user type a path, because when you type a
character into the field AND when you type return into the file dialog
(if you browse for the file) you get exactly the same event code.  It
may be necessary to use a different widget for typing file paths.