Add GitHub CSS style for the HTML files we create for installers
The `.css` file came from [https://gist.github.com/andyferra/2554919](https://gist.github.com/andyferra/2554919).  There was no license information or terms of distribution for the file in the GitHub Gist.

I created the `.html` in order to be able to use it in an argument to [pandoc](http://pandoc.org) in the `build.xml` file for the SBML Test Runner GUI.